Refactor DB test code
This factors the RealDatabaseMixin and fakedb code to allow test classes
to specify test data that should be inserted into the database before
the test is run.  This accomplishes two things: first, rows can be
inserted in proper dependency order, which is important for Postgres.
Second, it opens the possibility of using the same test scripts (e.g.,
test_db_changes.py) to test the fake database connectors used in testing
the higher-level parts of Buildbot.

Database cleanup is also improved, at the cost of making the database
test depend on the accuracy of buildbot.db.model.